+++ CHANGELOG.md
@@ -261,6 +261,12 @@
 - Removed the `neoblocks_owners` table; the Craft 5 `elements_owners` table is used instead
 - Removed the `neoblocks.deletedWithOwner` column; the Craft 5 `elements.deletedWithOwner` column is used instead
 
+## Unreleased
+
+### Added
+- Added `benf\neo\Field::$newBlockButtonLabel` (field setting for defining the block creation button text, on menu styles where it is shown)
+- Added `benf\neo\Field::newBlockButtonLabelWithFallback()`
+
 ## 4.2.26 - 2025-02-18
 
 ### Fixed